requirements.txt
openai
numpy
sounddevice
scipy
pydub
soundfile
elevenlabs
python-dotenv